// src/components/maps-view/enums/map-key.enum.ts
export enum MapKey {
  Tutorial = 'Tutorial',
  Barb_M1 = 'Barb_M1',
  Ragnar_M2 = 'Ragnar_M2',
  Dwarf_M3 = 'Dwarf_M3',
  Ruins_M4 = 'Ruins_M4',
  Mine_M5 = 'Mine_M5',
  Labyrinth_M6 = 'Labyrinth_M6',
  Tomb_M7 = 'Tomb_M7',
  Island_M8 = 'Island_M8',
  Orc_M9 = 'Orc_M9',
  Orlok_M10 = 'Orlok_M10',
  Ice_M11 = 'Ice_M11',
  Btomb_M12 = 'Btomb_M12',
  Desert_M13 = 'Desert_M13',
  Volcano_M14 = 'Volcano_M14',
  Palace_M15 = 'Palace_M15',
  Tower_M16 = 'Tower_M16',
  Chaos_M17 = 'Chaos_M17',
}